Added code behind Tumbly Fumbles
Note: Currently the method behind changing game state is designed as a toggle (toggle game timer for pause, for example). This was an experimental design choice and has led to bugs in the game. A more effective choice would be to decouple the notions of turning an aspect of the game on and turning it off.